Eliminate the need for B_OP_ALPHA in most cases.
If the colorspace is B_RGB32, then we don't need compositing.
But we also shouldn't need it in the case where we don't have a parent.

Only one additional redraw glitch added here (an intermittent single-pixel
black stripe in one context menu in Gecko.) All other applications
seem to work as before, including popup menus in GNOME Web, and the
buffer copying in app_server will be much more efficient this way.