Wait for Views to exit before deleting them.
Since WaylandView unlocks its looper and acquires the global lock
before Draw(), we can wind up in a situation where we are trying to
remove and delete the View while it's actually still in use.

In order to mitigate this, WaylandEnv now has a mechanism where it sets
a pointer to itself before unlocking the handler's looper, and unsets it
and optionally releases a semaphore in its destructor. This allows us
to wait for the view to finish.

Views can be attached in trees, though, so when removing a view with
children, we need to wait for all the descendants, too. A new method
RemoveDescendantsAndSelf() is introduced in WaylandView that takes
care of all this.

We now require that only this new method be the one to actually remove
views from their windows, so DetachedFromWindow() and its unlock-lock
dance is now gone, and replaced with an assertion in the destructor
that we were detached properly by the Surface. This fixes some more
races and use-after-frees.

Finally, HaikuXdgPopup and HaikuXdgToplevel are adjusted to detach
surfaces before destroying their parents, which brings things into
compliance with the previous requirement.

Fixes #9.